requirements.txt
asgiref==3.8.1
coverage==7.9.1
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
factory_boy==3.3.3
Faker==37.4.0
iniconfig==2.1.0
model-bakery==1.20.5
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.1
pytest==8.4.0
pytest-cov==6.2.1
pytest-django==4.11.1
sqlparse==0.5.3
tzdata==2025.2
